pyPrinter: fix comment rendering
The following
/*"hey"*/
is transformed to
""""hey""""
which is a syntax error. This commit fixes this edge case by trying to
use triple single quotes instead. If there are both a single and double
quotes at the beginning or end of the comment, we fallback to puttin the
comment on a new line.

This implementation still lacks the escaping of """ and ''' inside the
comment body.